Backend/requirements.txt
chromadb==0.6.3 
tiktoken==0.8.0
langchain==0.3.20
pydantic==2.10.4
pandas==2.1.4
huggingface-hub==0.29.2
torch==2.4.1
sentence-transformers==3.4.1
InstructorEmbedding==1.0.0
transformers==4.49.0
fastAPI==0.115.10
google-genai==1.9.0
uvicorn==0.34.0
pydantic-settings==2.7.1
email-validator==2.2.0
passlib==1.7.4
motor==3.7.0
langchain-community==0.3.19
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
scikit-learn==1.6.0
numpy==1.26.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
httpx>=0.27.0